Check Port to Determine TLS Use
Currently, the code only sets the pssl pointer when "https://" is explicitly specified. It might make sense for pssl to also be set to 1 when the port specified is 443, the standard HTTPS port. If this isn't in line with standard procedure, feel free to discard it; I figured it would just be something a developer would expect from the function.

CLA: trivial