EVP: Add internal functions to get relevant X509_ALGOR
Getting a relevant X509_ALGOR (AlgorithmIdentifier) for an EVP
operation or key it becoming central for production of diverse ASN.1
structures.  With EVP_PKEY_ASN1_METHODs, this was done with diverse
method functions that were more or less specialised for the task, and
that were also free to do any hackery they want.

With providers, we're avoiding the possibility for random hackery, but
still want to retain the possibility for provider code to produce the
necessary X509_ALGOR.

The mechanism for doing this is to ask the provider for the DER
encoded AlgorithmIdentifier instance for the operation or for the key,
determined by if it's asked for via an EVP_PKEY, an EVP_PKEY_CTX, an
EVP_MD, an EVP_MD_CTX, and so on, and use d2i_X509_ALGOR() with the
result to get the corresponding X509_ALGOR.

This feature was already introduced in an earlier commit with the
title "PROV: add RSA signature implementation", but without explaining
the mechanism for getting an X509_ALGOR.