CMS: Adapt for provider side keys
This involves being able to produce relevant X509_ALGORs.  However,
since CMS has its own set of standards for proper X509_ALGOR content,
which may or may not differ from the X.509 standards, we need to use a
slightly different parameter name for CMS specific use.  The CMS code
also tries the "normal" parameter name, for the cases where the
response would be the same for CMS as for X.509.