Fix PEM certificate loading that sometimes fails
As described in https://github.com/openssl/openssl/issues/9187 the
loading of PEM certificates sometimes fails if the line base64
content is in one line and the length of the line is a multiple of
254.
The problem is in get_header_and_data(). When the last line has a
length of 254 (or a multiple) the next read will only read a newline.
Due to this get_header_and_data() expects to be in the header not in
the data area. This commit fixes this by introducing the new state
MAYBE_POST_HEADER. This state ensures that the loaded content is
treated as data, if such a newline is the last thing that could be
read from the string. If there is additional content after the content
already read will be treated as data.