Expand CVE-2020-1967 tests to include client certificates
Although it is rare for a client to have multiple certificates it might
send, and thus need to call SSL_check_chain() to determine whether one
or the other is acceptable given the server's advertised parameters, it is
allowed by the library and protocol for the client to perform this operation.
As such, a client is still susceptible to a (with the current ecosystem,
malicious) server that sends unrecognized signature_algorithms_cert values
and NULL dereference within SSL_check_chain().

Add additional TLSProxy tests to test_sslsigalgs that inject (respectively)
"signature_algorithms" and "signature_algorithms_cert" extensions contiaining
unallocated codepoints.