Allow TLSProxy to parse empty Certificate messages
The client is allowed to send an empty Certificate message in response to a
CertificateRequest if the client does not have an appropriate certificate
available.  None of our current TLSProxy tests exercise this behavior
(or, it seems, client certificates at all), and so the parsing routine in
the TLSProxy attempted to read the length of an individual certificate from
a zero-length string and then operate on the resulting undefined value.

Recognize a zero-length TLS 1.3 Certificate message and skip over the rest
of the processing, just printing a note that an empty Certificate message
was received.