Warn that objects with opaque types must never be passed on.
If libcrypto passes an opaque object pointer to a function of a
provider that's linked with a different libcrypto, there's no saying
how that will work, as an opaque object may change at any time, even
between PATCH level.  Therefore, such objects must only be used as
arguments to upcalls, which are guaranteed to exist in the same
libcrypto that the objects came from.