Correct alignment calculation in ssl3_setup_write
The alignment calculation in ssl3_setup_write incorrectly results in a
value of 3. The maximum alignment actually used in do_ssl3_write is 7
bytes - so there is a potential to overrun the buffer by up to 4 bytes.

Fortunately, the encryption overhead allowed for is 80 bytes which
consists of 16 bytes for the cipher block size and 64 bytes for the MAC
output. However the biggest MAC that we ever produce is HMAC-384 which is
48 bytes - so we have a headroom of 16 bytes (i.e. more than the 4 bytes
of potential overrun).

Thanks to Nagesh Hegde for reporting this.

Fixes #11766