Configure: Add -Wno-unneeded-internal-declaration to --strict-warnings
This is for clang only.  The problem encountered is that some arrays
were deemed unnecessary by clang, for example:

    providers/common/der/der_rsa.c:424:28: error: variable 'der_aid_sha224Identifier' is not needed and will not be emitted [-Werror,-Wunneeded-internal-declaration]
    static const unsigned char der_aid_sha224Identifier[] = {
                               ^

However, these arrays are used in sizeof() expressions in other parts
of the code that's actually used, making that warning-turned-error not
quite accurate.

Better to turn it off in our --strict-warnings.