Enable legacy AES API to use AESNI or VPAES if available
When no assembler support is available, we fall back
to either the constant time C implementation or the
non-constant time C code implementation.

This is controlled by -DOPENSSL_AES_CONST_TIME.
So this makes the legacy API completely constant time,
if OPENSSL_AES_CONST_TIME is defined, otherwise it uses constant time
assembler implementations when available, and may fall back to the
non-constant time implementation.

This works so far only for intel and aarch64 CPUs.

[extended tests]